# openapi.yaml
openapi: 3.0.3
info:
  title: Hierarchical Agent System API
  description: |
    REST API for managing hierarchical agent teams, configurations, and task executions.
    
    This API provides comprehensive functionality for:
    - Creating and managing hierarchical agent teams
    - Discovering and searching agent libraries
    - Validating and optimizing team configurations
    - Executing tasks with real-time monitoring
    - Performance analytics and team evaluation
    
    ## Authentication
    Currently operates without authentication for development. Production deployments support:
    - API Key authentication via `X-API-Key` header
    - JWT Bearer tokens via `Authorization` header
    
    ## Rate Limiting
    - Default: 100 requests/minute per IP
    - Authenticated: 1000 requests/minute per API key
    - Execution endpoints: 10 concurrent executions per team
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details format with structured error responses.
    
  version: 1.0.0
  contact:
    name: Hierarchical Agents Support
    url: https://github.com/your-org/hierarchical-agents
    email: support@hierarchical-agents.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.hierarchical-agents.com
    description: Production server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - {}
...
paths:
...
  /api/v1/agents/search:
    post:
      tags: [agents]
      summary: Search Agents
      description: Search for agents based on capabilities, role, or other criteria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentSearchRequest'
            examples:
              capability_search:
                summary: Search by Capabilities
                value:
                  capabilities: ["research", "analysis"]
                  limit: 10
              role_search:
                summary: Search by Role
                value:
                  role: "worker"
                  domain: "research"
                  min_compatibility: 0.8
      responses:
        '200':
          description: List of matching agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
  /api/v1/agents/:
    get:
      tags: [agents]
      summary: List Agents
      description: List all available agents with filtering
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [coordinator, supervisor, worker, specialist]
        - name: capability
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
  /api/v1/agents/{agent_id}:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    
    get:
      tags: [agents]
      summary: Get Agent Details
      description: Get detailed information about a specific agent
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/agents/compatibility:
    post:
      tags: [agents]
      summary: Check Agent Compatibility
      description: Check compatibility between multiple agents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCompatibilityRequest'
      responses:
        '200':
          description: Compatibility analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentCompatibilityResponse'
...
  /api/v1/configs/validate:
    post:
      tags: [configs]
      summary: Validate Configuration
      description: Validate a hierarchical team configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigValidationRequest'
      responses:
        '200':
          description: Validation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigValidationResponse'
  /api/v1/configs/upload:
    post:
      tags: [configs]
      summary: Upload Configuration File
      description: Upload and parse a configuration file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: YAML or JSON configuration file
                validate:
                  type: boolean
                  default: true
                  description: Auto-validate after upload
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigUploadResponse'
  /api/v1/configs/templates:
    get:
      tags: [configs]
      summary: List Templates
      description: List available configuration templates
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, popularity, rating, created_at]
            default: name
      responses:
        '200':
          description: List of templates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateListResponse'
  /api/v1/configs/templates/{template_id}:
    parameters:
      - $ref: '#/components/parameters/TemplateId'
    
    get:
      tags: [configs]
      summary: Get Template Details
      description: Get detailed template information and configuration
      responses:
        '200':
          description: Template details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/configs/export:
    post:
      tags: [configs]
      summary: Export Configuration
      description: Export team configuration in various formats
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigExportRequest'
      responses:
        '200':
          description: Exported configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigExportResponse'

  # Task Execution
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AgentId:
      name: agent_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_-]+$'
      description: Unique agent identifier
      example: agent_researcher_001
...
    TemplateId:
      name: template_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_-]+$'
      description: Unique template identifier
      example: template_research_pipeline
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "NOT_FOUND"
              message: "Team not found"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Request validation failed"
            details:
              type: object
              description: Additional error context
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
              example: "req_abc123"

    # Team Models
...
    TeamConfiguration:
      type: object
      required: [coordinator, teams]
      properties:
        coordinator:
          $ref: '#/components/schemas/CoordinatorConfig'
        teams:
          type: array
          items:
            $ref: '#/components/schemas/TeamConfig'
          minItems: 1
        performance:
          $ref: '#/components/schemas/PerformanceConfig'
    CoordinatorConfig:
      type: object
      required: [name, description, llm, prompts]
      properties:
        name:
          type: string
          example: "Research Coordinator"
        description:
          type: string
          example: "Coordinates research tasks across specialized teams"
        llm:
          $ref: '#/components/schemas/LLMConfig'
        prompts:
          $ref: '#/components/schemas/CoordinatorPromptsConfig'
        routing:
          $ref: '#/components/schemas/RoutingConfig'
        memory:
          $ref: '#/components/schemas/MemoryConfig'
    CoordinatorPromptsConfig:
      type: object
      required: [system_prompt, decision_prompt]
      properties:
        system_prompt:
          $ref: '#/components/schemas/PromptTemplate'
        decision_prompt:
          $ref: '#/components/schemas/PromptTemplate'
        coordination_prompt:
          $ref: '#/components/schemas/PromptTemplate'
    RoutingConfig:
      type: object
      properties:
        strategy:
          type: string
          enum: [keyword_based, llm_based, rule_based, hybrid]
          default: hybrid
        fallback_team:
          type: string
        fallback_worker:
          type: string
        confidence_threshold:
          type: number
          default: 0.7
        max_decision_time:
          type: integer
          default: 30
    TeamConfig:
      type: object
      required: [name, supervisor, workers]
      properties:
        name:
          type: string
          example: "Research Team"
        supervisor:
          $ref: '#/components/schemas/SupervisorConfig'
        workers:
          type: array
          items:
            $ref: '#/components/schemas/WorkerConfig'
          minItems: 1
    SupervisorConfig:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Research Supervisor"
        prompts:
          $ref: '#/components/schemas/PromptConfig'
        llm:
          $ref: '#/components/schemas/LLMConfig'
        memory:
          $ref: '#/components/schemas/MemoryConfig'
    WorkerConfig:
      type: object
      required: [name, capabilities]
      properties:
        name:
          type: string
          example: "Senior Researcher"
        capabilities:
          type: array
          items:
            type: string
          example: ["research", "analysis", "data_processing"]
        tools:
          type: array
          items:
            type: string
          example: ["web_search", "calculator", "file_reader"]
        specializations:
          type: array
          items:
            type: string
          example: ["academic_research", "market_analysis"]
        config_file:
          type: string
          example: "configs/examples/senior_researcher.yml"
        config_data:
          type: object
          description: Inline configuration data
    LLMConfig:
      type: object
      required: [provider, model]
      properties:
        provider:
          type: string
          enum: [openai, anthropic, google, groq]
        model:
          type: string
          example: "gpt-4o"
        temperature:
          type: number
          minimum: 0
          maximum: 2
          example: 0.7
        max_tokens:
          type: integer
          minimum: 1
          example: 2048
        timeout:
          type: integer
          minimum: 1
          example: 30
        api_key_env:
          type: string
          example: "OPENAI_API_KEY"
          description: "Environment variable name containing the API key"
    PromptTemplate:
      type: object
      required: [template]
      properties:
        template:
          type: string
          example: "You are a coordinator managing specialized teams..."
        variables:
          type: array
          items:
            type: string
          example: ["task_context", "team_status"]
    PromptConfig:
      type: object
      properties:
        system_prompt:
          $ref: '#/components/schemas/PromptTemplate'
    MemoryConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        type:
          type: string
          enum: [semantic, episodic, procedural]
        retention_policy:
          type: string
          enum: [session, persistent, limited]
    PerformanceConfig:
      type: object
      properties:
        monitoring_enabled:
          type: boolean
          default: true
        alerts:
          type: array
          items:
            type: object
            properties:
              metric:
                type: string
              threshold:
                type: number
              action:
                type: string

    # Agent Models
    AgentSearchRequest:
      type: object
      properties:
        query:
          type: string
          maxLength: 200
          example: "research analysis"
        role:
          type: string
          enum: [coordinator, supervisor, worker, specialist]
        capabilities:
          type: array
          items:
            type: string
          example: ["research", "analysis"]
        tools:
          type: array
          items:
            type: string
        domain:
          type: string
          example: "academic"
        min_compatibility:
          type: number
          minimum: 0
          maximum: 1
          example: 0.8
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        offset:
          type: integer
          minimum: 0
          default: 0
    AgentListResponse:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: '#/components/schemas/AgentResponse'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    AgentResponse:
      type: object
      properties:
        id:
          type: string
          example: "agent_researcher_001"
        name:
          type: string
          example: "Senior Researcher"
        description:
          type: string
          example: "Expert in academic research and data analysis"
        primary_role:
          type: string
          enum: [coordinator, supervisor, worker, specialist]
        secondary_roles:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            $ref: '#/components/schemas/Capability'
        tools:
          type: array
          items:
            type: string
        specializations:
          type: array
          items:
            type: string
        compatibility_score:
          type: number
          minimum: 0
          maximum: 1
        file_path:
          type: string
        can_coordinate:
          type: boolean
        can_supervise:
          type: boolean
        team_size_limit:
          type: integer
    Capability:
      type: object
      properties:
        name:
          type: string
          example: "research"
        description:
          type: string
          example: "Academic and web research"
        proficiency:
          type: number
          minimum: 0
          maximum: 1
          example: 0.9
    AgentCompatibilityRequest:
      type: object
      required: [agent_ids]
      properties:
        agent_ids:
          type: array
          items:
            type: string
          minItems: 2
          maxItems: 20
        team_context:
          type: string
        include_suggestions:
          type: boolean
          default: true
    AgentCompatibilityResponse:
      type: object
      properties:
        compatibility_matrix:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: number
        overall_compatibility:
          type: number
        recommendations:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
...
    ConfigValidationRequest:
      type: object
      required: [config]
      properties:
        config:
          $ref: '#/components/schemas/TeamConfiguration'
        validation_level:
          type: string
          enum: [basic, standard, strict]
          default: standard
        optimize:
          type: boolean
          default: false
    ConfigValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
        validation_result:
          type: object
          properties:
            overall_score:
              type: number
            issues:
              type: array
              items:
                $ref: '#/components/schemas/ValidationIssue'
            warnings:
              type: array
              items:
                type: string
            recommendations:
              type: array
              items:
                type: string
            optimized_config:
              $ref: '#/components/schemas/TeamConfiguration'
    ValidationIssue:
      type: object
      properties:
        severity:
          type: string
          enum: [error, warning, info]
        code:
          type: string
        message:
          type: string
        path:
          type: string
        suggestion:
          type: string
    ConfigUploadResponse:
      type: object
      properties:
        file_id:
          type: string
        filename:
          type: string
        size:
          type: integer
        format:
          type: string
          enum: [yaml, json]
        config:
          $ref: '#/components/schemas/TeamConfiguration'
        validation_result:
          $ref: '#/components/schemas/ConfigValidationResponse'
    TemplateListResponse:
      type: object
      properties:
        templates:
          type: array
          items:
            $ref: '#/components/schemas/TemplateSummary'
        total:
          type: integer
    TemplateSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        tags:
          type: array
          items:
            type: string
        usage_count:
          type: integer
        rating:
          type: number
        created_at:
          type: string
          format: date-time
        file_path:
          type: string
    TemplateResponse:
      allOf:
        - $ref: '#/components/schemas/TemplateSummary'
        - type: object
          properties:
            config:
              $ref: '#/components/schemas/TeamConfiguration'
            examples:
              type: array
              items:
                type: object
            customization_options:
              type: object
            prerequisites:
              type: array
              items:
                type: string
    ConfigExportRequest:
      type: object
      required: [team_id, format]
      properties:
        team_id:
          type: string
        format:
          type: string
          enum: [yaml, json, docker, terraform]
        include_metadata:
          type: boolean
          default: true
        anonymize:
          type: boolean
          default: false
    ConfigExportResponse:
      type: object
      properties:
        format:
          type: string
        content:
          type: string
        filename:
          type: string
        size:
          type: integer

    # Execution Models
...